Move Espoo BI stuff under espoo package
