+++ service/src/main/kotlin/fi/espoo/evaka/espoo/bi/EspooBiModels.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-package fi.espoo.evaka.espoobi
+package fi.espoo.evaka.espoo.bi
 
 import fi.espoo.evaka.application.ApplicationOrigin
 import fi.espoo.evaka.application.ApplicationStatus
